Fix test of average error/deviation
When computing the average error/deviation from the reference data, the
computed value may be negative. The error bound it always a positive
number, therefore we ensure that the computed values is also positive.

This requires slight increase in an error bound of the timeshift. No
deep investigation has been done regarding the correctness because the
increase is small and the timeshift code is currently not used to our
knowledge.