Add initial version of impedance Python bindings
The way the Python bindings are packaged are inspired by the Scientific
Python guide [1] and pybind's CMake example [4]. scikit-build-core [2]
is chosen as it promises to be a modern rewrtie of scikit-build, is part
of the Scientific Python examples and offers all features needed for
creating the Python bindings and is also the preferred build system in
the pybind11 [3] documentation.

scikit-build-core works best with CMake 3.15 and later. We therefore
bump the required CMake version.

pybind11 [5] is used for the bindings as it appeared simple to create
bindings and because it was recommended by colleagues. We don't expose
the created bindings directly to the users, but wrap them in another
Python function. This allows us to create take xtgeo.Cube objects as
input and output without handling this type in the actual bindings. Only
NumPy arrays are passed between Python and C++ for which pybind11 has
native support. xtgeo.Cube are the preferred input and output data types
for the anticipated workflows. For simplicity, pybind11 was added as git
submodule. This reduces friction when building the Python bindings from
scratch since it is straightforward for CMake to find pybind11.

[1]: https://learn.scientific-python.org/development/
[2]: https://scikit-build-core.readthedocs.io/en/latest/
[3]: https://pybind11.readthedocs.io/en/stable/compiling.html
[4]: https://github.com/pybind/scikit_build_example
[5]: https://pybind11.readthedocs.io/en/stable/index.html