Remove Travis CI configuration
Travis CI is not used anymore but GitHub Actions. With the recent
changes the Travis CI workflow is reproduced by GitHub Actions.
Therefore, we do not need to keep the outdated configuration files
anymore.

New releases on GitHub must be created manually. The pre-compiled
command line tools need to be downloaded and then added manually to the
release. The Python wheels are available on PyPI and are uploaded there
automatically. Further automation of releases is possible, but we
currently do not expect many releases to justify the creation and
maintenance of such a job.